Add mining system and expand movement action capabilities
- Implement block mining with break time calculation based on hardness and tool
- Add visible block detection with efficient surface raycasting
- Improve action execution with path return and separate execution step
- Add comprehensive agent logging to file and stdout
- Enhance logging infrastructure for debugging movement and actions
- Improve type safety across action and movement systems

+++ agent/items_adapters.go
@@ -196,7 +196,7 @@ type registryItemManager struct {
 	registryGetter func(string) models.CustomRegistry
 }
 
-func (m registryItemManager) GetItemNameByID(id int) string {
+func (m registryItemManager) GetItemNameByID(id int32) string {
 	if m.registryGetter == nil {
 		return ""
 	}
@@ -204,7 +204,7 @@ func (m registryItemManager) GetItemNameByID(id int) string {
 	if reg == nil || !reg.IsReady() {
 		return ""
 	}
-	if name, ok := reg.GetNameByID(int32(id)); ok {
+	if name, ok := reg.GetNameByID(id); ok {
 		return name
 	}
 	return ""
